EL-103 Список уведомлений пациента у врача и родителя

+++ feature/analytics/src/main/java/ru/dekabrsky/analytics/AnalyticsSender.kt
@@ -25,7 +25,7 @@ class AnalyticsSender @Inject constructor(
         firebaseAnalytics.logEvent(name, bundle)
     }
 
-    fun setUserId(id: Int) {
+    fun setUserId(id: Long) {
         firebaseAnalytics.setUserId(id.toString())
     }
