+++ feature/chats/src/main/java/ru/dekabrsky/callersbase/domain/model/ChatEntity.kt
@@ -4,15 +4,15 @@ import org.threeten.bp.LocalDateTime
 import ru.dekabrsky.feature.loginCommon.domain.model.UserInfoEntity
 
 class ChatEntity(
-    val chatId: Int,
+    val chatId: Long,
     val firstUser: UserInfoEntity,
     val secondUser: UserInfoEntity,
     val messages: List<MessageEntity>
 )
 
 class MessageEntity(
-    val messageId: Int = 0,
-    val userId: Int = 0,
+    val messageId: Long = 0,
+    val userId: Long = 0,
     val text: String = "",
     val createdDate: LocalDateTime = LocalDateTime.now()
 )
