+++ feature/chats/src/main/java/ru/dekabrsky/callersbase/presentation/model/ChatConversationScreenArgs.kt
@@ -4,6 +4,6 @@ import ru.dekabrsky.feature.loginCommon.domain.model.UserInfoEntity
 import java.io.Serializable
 
 class ChatConversationScreenArgs(
-    val chatId: Int,
+    val chatId: Long,
     val companion: UserInfoEntity
 ): Serializable
