+++ feature/chats/src/main/java/ru/dekabrsky/callersbase/presentation/model/ChatFlowCache.kt
@@ -3,5 +3,5 @@ package ru.dekabrsky.callersbase.presentation.model
 import javax.inject.Inject
 
 class ChatFlowCache @Inject constructor() {
-    var existingCompanionIds: List<Int> = listOf()
+    var existingCompanionIds: List<Long> = listOf()
 }
