+++ feature/events-common/src/main/java/ru/dekabrsky/common/domain/model/DialingEntity.kt
@@ -7,7 +7,7 @@ class DialingEntity (
     val id: Int,
     val name: String,
 
-    val callersBaseId: Int,
+    val callersBaseId: Long,
     val callersBaseName: String,
     val callersBaseCount: Int,
